+++ android/app/src/main/kotlin/com/jne_solutions/jne_household_app/glance/TotalBudgetReceiver.kt
@@ -1,0 +1,7 @@
+package com.jne_solutions.jne_household_app.glance
+
+import HomeWidgetGlanceWidgetReceiver
+
+class TotalBudgetReceiver : HomeWidgetGlanceWidgetReceiver<TotalBudgetWidget>() {
+    override val glanceAppWidget = TotalBudgetWidget()
+}