+++ lib/widgets_shared/dialogs/bank_account_dialog.dart
@@ -1,6 +1,6 @@
 
 import 'package:flutter/material.dart';
-import 'package:jne_household_app/helper/text_formatter.dart';
+import 'package:jne_household_app/services/text_formatter.dart';
 import 'package:jne_household_app/i18n/i18n.dart';
 import 'package:jne_household_app/models/bankaccount.dart';
 import 'package:jne_household_app/models/booking_principles.dart';