+++ lib/widgets_shared/home/budget_summary.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:jne_household_app/helper/debug_screenshot_manager.dart';
+import 'package:jne_household_app/services/debug_screenshot_manager.dart';
 import 'package:jne_household_app/i18n/i18n.dart';
 import 'package:jne_household_app/models/design_state.dart';
 import 'package:jne_household_app/widgets_shared/home/budget_arc.dart';
@@ -29,6 +29,8 @@ class _BudgetSummaryState extends State<BudgetSummary> {
     final currency = budgetState.currency;
     final List<double> categorySpent = budgetState.categories.map((c) => c.spent).toList();
     final double totalSpent = categorySpent.fold(0.0, (sum, spent) => sum + spent);
+
+    budgetState.updateTotalSpentWidget(totalSpent);
 
     if (kDebugMode && !Platform.isAndroid && !Platform.isIOS) {
       ScreenshotManager().takeScreenshot(name: "main");