+++ lib/widgets_shared/home/category_list.dart
@@ -4,8 +4,9 @@ import 'dart:ui';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
-import 'package:jne_household_app/helper/brightness.dart';
-import 'package:jne_household_app/helper/debug_screenshot_manager.dart';
+import 'package:jne_household_app/screens_mobile/mobile_receipt_scanner.dart';
+import 'package:jne_household_app/services/brightness.dart';
+import 'package:jne_household_app/services/debug_screenshot_manager.dart';
 import 'package:jne_household_app/models/budget_state.dart';
 import 'package:jne_household_app/models/category_budget.dart';
 import 'package:jne_household_app/models/design_state.dart';
@@ -87,7 +88,7 @@ Widget categoryList(String currency, BudgetState budgetState, BuildContext conte
                 : 
                 (category.spent <= budgetState.notAssignedBudget))
                 ? 
-                  ((designState.categoryMainStyle == 0) ? getTextColor(category.color.withAlpha((allSpent) ? 51 : 255), designState.categoryMainStyle, context) : Theme.of(context).textTheme.bodyMedium!.color!)
+                  ((designState.categoryMainStyle == 0) ? getTextColor(category.color.withAlpha((allSpent) ? 51 : 255), designState.categoryMainStyle, context: context) : Theme.of(context).textTheme.bodyMedium!.color!)
                   : 
                   Colors.red
                 );
@@ -109,14 +110,24 @@ Widget categoryList(String currency, BudgetState budgetState, BuildContext conte
             category.color,
             designState
           ),
-          onPressed: () => showExpenseDialog(
-            context: context,
-            category: category.category,
-            categoryId: category.categoryId,
-            accountId: budgetState.filterBudget,
-            bankAccounts: budgetState.bankAccounts,
-            bankAccoutCount: budgetState.bankAccounts.length
-          )
+          onPressed: () async {
+            bool res = await showExpenseDialog(
+              context: context,
+              category: category.category,
+              categoryId: category.categoryId,
+              accountId: budgetState.filterBudget,
+              bankAccounts: budgetState.bankAccounts,
+              bankAccoutCount: budgetState.bankAccounts.length,
+              allowCamera: (kDebugMode || budgetState.isPro) && (Platform.isAndroid || Platform.isIOS)
+            );
+            if (res) {
+              Navigator.of(context).push(
+                MaterialPageRoute(
+                  builder: (context) => ReceiptPage(baseCurrency: budgetState.currency, budgetState: budgetState, designState: designState, overrideCatId: category.categoryId, closeAfterSuccess: true),
+                ),
+              );
+            }
+          }
         );
       },
     );
@@ -131,7 +142,7 @@ Widget categoryList(String currency, BudgetState budgetState, BuildContext conte
   }
 }
 
-Widget listTile({required context, required bool allSpent, required bool unassigned, required CategoryBudget category, required Color textColor, required BudgetState budgetState, required String currency, required DesignState designState, required Function buttonBuilder, required VoidCallback showExpensesBottomSheet, required VoidCallback onPressed}) {
+Widget listTile({required context, required bool allSpent, required bool unassigned, required CategoryBudget category, required Color textColor, required BudgetState budgetState, required String currency, required DesignState designState, required Function buttonBuilder, required VoidCallback showExpensesBottomSheet, required VoidCallback onPressed}) {  
   return Card(
     shape: RoundedRectangleBorder(
       borderRadius: BorderRadius.circular(8),
@@ -141,48 +152,51 @@ Widget listTile({required context, required bool allSpent, required bool unassig
     children: [ 
       Stack(
         children: [
-            Positioned(
-              height: (designState.categoryMainStyle == 0) ? null : 5,
-              left: 0,
-              right: 0,
-              top: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 1) ? null : 0,
-              bottom: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 2) ? null : 0,
-              child: Container(
-                decoration: BoxDecoration(
-                  color: category.color.withValues(alpha: (allSpent)? .2 : 1),
-                  borderRadius: BorderRadius.circular(8),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withValues(alpha: 0.4),
-                      blurRadius: 8,
-                      offset: const Offset(0, 4),
-                    )
-                  ],
+            if (designState.categoryMainStyle != 3)
+            ...[
+              Positioned(
+                height: (designState.categoryMainStyle == 0) ? null : 5,
+                left: 0,
+                right: 0,
+                top: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 1) ? null : 0,
+                bottom: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 2) ? null : 0,
+                child: Container(
+                  decoration: BoxDecoration(
+                    color: category.color.withValues(alpha: (allSpent)? .2 : 1),
+                    borderRadius: BorderRadius.circular(8),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.black.withValues(alpha: 0.4),
+                        blurRadius: 8,
+                        offset: const Offset(0, 4),
+                      )
+                    ],
+                  ),
                 ),
               ),
-            ),
-            Positioned(
-              height: (designState.categoryMainStyle == 0) ? null : 5,
-              left: 0,
-              right: 0,
-              top: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 1) ? null : 0,
-              bottom: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 2) ? null : 0,
-              child: Align(
-                alignment: Alignment.centerRight,
-                child: FractionallySizedBox(
-                  widthFactor: (!unassigned) ? (category.spent / category.budget).clamp(0.0, 1.0) : (category.spent / budgetState.notAssignedBudget).clamp(0.0, 1.0),
-                  child: Container(
-                    decoration: BoxDecoration(
-                      color: Colors.black.withValues(alpha: (allSpent)? 0 : 0.5),
-                      borderRadius: const BorderRadius.horizontal(
-                        left: Radius.circular(8),
-                        right: Radius.circular(8),
+              Positioned(
+                height: (designState.categoryMainStyle == 0) ? null : 5,
+                left: 0,
+                right: 0,
+                top: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 1) ? null : 0,
+                bottom: (designState.categoryMainStyle == 0) ? 0 : (designState.categoryMainStyle == 2) ? null : 0,
+                child: Align(
+                  alignment: Alignment.centerRight,
+                  child: FractionallySizedBox(
+                    widthFactor: (!unassigned) ? (category.spent / category.budget).clamp(0.0, 1.0) : (category.spent / budgetState.notAssignedBudget).clamp(0.0, 1.0),
+                    child: Container(
+                      decoration: BoxDecoration(
+                        color: Colors.black.withValues(alpha: (allSpent)? 0 : 0.5),
+                        borderRadius: const BorderRadius.horizontal(
+                          left: Radius.circular(8),
+                          right: Radius.circular(8),
+                        ),
                       ),
                     ),
                   ),
                 ),
               ),
-            ),
+            ],
             ListTile(
               title: Text(
                 (!unassigned)
@@ -223,6 +237,15 @@ Widget listTile({required context, required bool allSpent, required bool unassig
                 label: I18n.translate("new")
               ),
               onTap: showExpensesBottomSheet,
+              onLongPress: () {
+                if ((Platform.isAndroid || Platform.isIOS) && (budgetState.isPro || kDebugMode)) {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (context) => ReceiptPage(baseCurrency: budgetState.currency, budgetState: budgetState, designState: designState, overrideCatId: category.categoryId, closeAfterSuccess: true),
+                    ),
+                  );
+                }
+              },
             )
           ]
         )